add devops tools & clean code
