+++ .eslintrc.js
@@ -1,0 +1,7 @@
+const rules = {
+}
+
+module.exports = {
+  extends: '@chatie',
+  rules,
+}
