+++ .markdownlintrc
@@ -1,0 +1,13 @@
+{
+  "default": true,
+  "no-trailing-punctuation": {
+    "punctuation": ".,;:!"
+  },
+  "MD013": false,
+  "MD033": false,
+  "first-line-h1": false,
+  "no-hard-tabs": true,
+  "no-trailing-spaces": {
+    "br_spaces": 2
+  }
+}
