+++ .vscode/settings.json
@@ -1,0 +1,98 @@
+{
+  "typescript.tsdk": "./node_modules/typescript/lib",
+  "editor.fontLigatures": true,
+
+  "editor.tokenColorCustomizations": {
+    "textMateRules": [
+      {
+        "scope": [
+          //following will be in italics (=Pacifico)
+          "comment",
+          // "entity.name.type.class", //class names
+          "keyword", //import, export, return…
+          "support.class.builtin.js", //String, Number, Boolean…, this, super
+          "storage.modifier", //static keyword
+          "storage.type.class.js", //class keyword
+          "storage.type.function.js", // function keyword
+          "storage.type.js", // Variable declarations
+          "keyword.control.import.js", // Imports
+          "keyword.control.from.js", // From-Keyword
+          "entity.name.type.js", // new … Expression
+          "keyword.control.flow.js", // await
+          "keyword.control.conditional.js", // if
+          "keyword.control.loop.js", // for
+          "keyword.operator.new.js", // new
+        ],
+        "settings": {
+          "fontStyle": "italic",
+        },
+      },
+      {
+        "scope": [
+          //following will be excluded from italics (My theme (Monokai dark) has some defaults I don't want to be in italics)
+          "invalid",
+          "keyword.operator",
+          "constant.numeric.css",
+          "keyword.other.unit.px.css",
+          "constant.numeric.decimal.js",
+          "constant.numeric.json",
+          "entity.name.type.class.js"
+        ],
+        "settings": {
+          "fontStyle": "",
+        },
+      }
+    ]
+  },
+
+  "files.exclude": {
+    "dist/": true,
+    "doc/": true,
+    "node_modules/": true,
+    "package/": true,
+  },
+  "alignment": {
+    "operatorPadding": "right",
+    "indentBase":      "firstline",
+    "surroundSpace":   {
+      "colon":      [1, 1],   // The first number specify how much space to add to the left, can be negative. The second number is how much space to the right, can be negative.
+      "assignment": [1, 1],    // The same as above.
+      "arrow":      [1, 1],    // The same as above.
+      "comment":    2,         // Special how much space to add between the trailing comment and the code.
+                               // If this value is negative, it means don't align the trailing comment.
+    }
+  },
+  "editor.formatOnSave": false,
+  "python.pythonPath": "python3",
+  "eslint.validate": [
+    "javascript",
+    "typescript",
+  ],
+  "cSpell.words": [
+    "Gitter",
+    "Wechaty",
+    "appid",
+    "botie",
+    "huan",
+    "ioscat",
+    "ipad",
+    "lcov",
+    "lijiarui",
+    "logonoff",
+    "padchat",
+    "padplus",
+    "padpro",
+    "pagepath",
+    "portfinder",
+    "qrcode",
+    "removee",
+    "ruirui",
+    "thumbnailurl",
+    "wechat",
+    "weixin",
+    "wxwork",
+    "zbeekman",
+    "zixia"
+  ],
+  "typescript.format.semicolons": "remove",
+}
