+++ package.json
@@ -2,11 +2,31 @@
   "name": "wechaty-cli",
   "version": "0.0.1",
   "description": "Command-line Client for Wechaty, powered by blessed",
-  "main": "index.ts",
-  "scripts": {
-    "start": "ts-node index.ts",
-    "test": "echo \"Error: no test specified\" && exit 1"
+  "main": "dist/src/mod.js",
+  "typings": "dist/src/mod.d.ts",
+  "engines": {
+    "wechaty": ">=0.60"
   },
+  "bin": {
+    "wechaty-cli": "dist/bin/cli.js"
+  },
+  "scripts": {
+    "clean": "shx rm -fr dist/*",
+    "dist": "npm run clean && tsc",
+    "pack": "npm pack",
+    "start": "ts-node examples/ding-dong-bot.ts",
+    "lint": "npm run lint:es && npm run lint:ts && npm run lint:md",
+    "lint:md": "markdownlint README.md",
+    "lint:ts": "tsc --noEmit",
+    "lint:es": "eslint \"src/**/*.ts\" \"tests/**/*.spec.ts\" --ignore-pattern tests/fixtures/",
+    "test": "npm run lint && npm run test:unit",
+    "test:pack": "bash -x scripts/npm-pack-testing.sh",
+    "test:unit": "blue-tape -r ts-node/register 'src/**/*.spec.ts' 'tests/**/*.spec.ts'"
+  },
+  "keywords": [
+    "wechaty",
+    "token"
+  ],
   "author": "chinggg",
   "license": "Apache-2.0",
   "dependencies": {
@@ -19,6 +39,24 @@
   "devDependencies": {
     "tsconfig-paths": "^3.10.1",
     "wechaty-puppet-wechat": "^0.28.3",
-    "wechaty-puppet-wechat4u": "^0.18.2"
+    "wechaty-puppet-wechat4u": "^0.18.2",
+    "@chatie/eslint-config": "^0.12.4",
+    "@chatie/git-scripts": "^0.6.2",
+    "@chatie/semver": "^0.4.7",
+    "@chatie/tsconfig": "^0.16.2",
+    "@grpc/grpc-js": "^1.3.6",
+    "@types/uuid": "^8.3.1",
+    "pkg-jq": "^0.2.11",
+    "shx": "^0.3.3",
+    "tstest": "^0.4.10"
+  },
+  "publishConfig": {
+    "access": "public",
+    "tag": "next"
+  },
+  "git": {
+    "scripts": {
+      "pre-push": "npx git-scripts-pre-push"
+    }
   }
 }
