+++ scripts/npm-pack-testing.sh
@@ -1,0 +1,36 @@
+#!/usr/bin/env bash
+set -e
+
+VERSION=$(npx pkg-jq -r .version)
+
+if npx --package @chatie/semver semver-is-prod "$VERSION"; then
+  NPM_TAG=latest
+else
+  NPM_TAG=next
+fi
+
+npm run dist
+npm run pack
+
+TMPDIR="/tmp/npm-pack-testing.$$"
+mkdir "$TMPDIR"
+mv ./*-*.*.*.tgz "$TMPDIR"
+cp tests/fixtures/smoke-testing.ts "$TMPDIR"
+
+cd $TMPDIR
+npm init -y
+npm install ./*-*.*.*.tgz \
+  @chatie/tsconfig \
+  @grpc/grpc-js \
+
+./node_modules/.bin/tsc \
+  --esModuleInterop \
+  --lib esnext \
+  --skipLibCheck \
+  --noEmitOnError \
+  --noImplicitAny \
+  --target es6 \
+  --module commonjs \
+  smoke-testing.ts
+
+node smoke-testing.js
