+++ scripts/package-publish-config-tag.sh
@@ -1,0 +1,13 @@
+#!/usr/bin/env bash
+set -e
+
+VERSION=$(npx pkg-jq -r .version)
+
+if npx --package @chatie/semver semver-is-prod $VERSION; then
+  npx pkg-jq -i '.publishConfig.tag="latest"'
+  echo "production release: publicConfig.tag set to latest."
+else
+  npx pkg-jq -i '.publishConfig.tag="next"'
+  echo 'development release: publicConfig.tag set to next.'
+fi
+
