+++ src/mod.ts
@@ -1,0 +1,5 @@
+import { VERSION } from './version'
+
+export {
+  VERSION,
+}
