+++ src/version.ts
@@ -1,0 +1,7 @@
+
+/**
+ * This file will be overwrite when we publish NPM module
+ * by scripts/generate_version.ts
+ */
+
+export const VERSION = '0.0.0'
