+++ tests/integration.spec.ts
@@ -1,0 +1,7 @@
+#!/usr/bin/env ts-node
+
+import { test }  from 'tstest'
+
+test('integration testing', async t => {
+  t.skip('tbw')
+})
